use uint storage
int storage is problematic for fv and should never be used anyway

+++ src/transferFrom.sol
@@ -18,7 +18,7 @@
 pragma solidity ^0.4.24;
 
 contract VatLike {
-    function dai(bytes32) public view returns (int);
+    function dai(bytes32) public view returns (uint);
     function Tab() public view returns (uint);
     function move(bytes32,bytes32,uint) public;
 }
@@ -30,7 +30,7 @@ contract Dai20 {
     uint256 constant ONE = 10 ** 27;
 
     function balanceOf(address guy) public view returns (uint) {
-        return uint(vat.dai(bytes32(guy))) / ONE;
+        return vat.dai(bytes32(guy)) / ONE;
     }
     function totalSupply() public view returns (uint) {
         return vat.Tab() / ONE;
